Refactor AppPy
